formated the variable block to look more like a forest
had this great variable block from which all this great hoisting and declaration comes from but it was just sitting there on the page being a single line. Reformatted the whitespace to make a forest path that the narration runs right through (note the width is more narrow than the paragraph so you get a nice feeling of being cramped as you move down the path).